# docs/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "The Electrum Bill Payments Service describes an interface for performing\
    \ bill payments\n                  transactions. These transactions require a\
    \ reference number which uniquely identifies the bill\n                  processor,\
    \ bill issuer, as well as the customer account, and is typically printed on a\
    \ customer\n                  invoice. The customer then makes use of this reference\
    \ number to make a payment towards the\n                  corresponding bill."
  version: "1.0.0"
  title: "Bill Payments Service Interface"
  contact:
    name: "Electrum Support"
    url: "http://io.electrum"
    email: "support@electrum.co.za"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "sandbox.electrum.io"
basePath: "/billpay/v1"
schemes:
- "https"
paths:
  /accountLookups:
    post:
      summary: "Request account info"
      description: "Requests the current status of the account identified by the supplied\
        \ account reference. Not all services support account lookups. In the case\
        \ where this function is not supported for the requested service, a 501 HTTP\
        \ status code may be returned"
      operationId: "requestAccountInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "An account lookup request"
        required: true
        schema:
          $ref: "#/definitions/AccountLookupRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/AccountLookupResponse"
        400:
          description: "Bad request"
          schema:
            $ref: "#/definitions/ErrorDetail"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/ErrorDetail"
        501:
          description: "Not implemented - in the case where the service does not support\
            \ account lookups"
        503:
          description: "Service Unavailable"
          schema:
            $ref: "#/definitions/ErrorDetail"
        504:
          description: "Gateway Timeout"
          schema:
            $ref: "#/definitions/ErrorDetail"
      security:
      - httpBasic: []
...
securityDefinitions:
  httpBasic:
    description: "All requests require HTTP basic authentication, where user name\
      \ equals the value used in\n                    MessageId.institutionId and\
      \ password has been agreed with Electrum."
    type: "basic"
definitions:
  Account:
    type: "object"
    required:
    - "accountRef"
    properties:
      accountRef:
        type: "string"
        description: "A reference number identifying the bill payments processor,\
          \ bill issuer, and customer"
        minLength: 6
        maxLength: 40
      balance:
        $ref: "#/definitions/LedgerAmount"
      dueDate:
        type: "string"
        format: "date-time"
        description: "The effective date of the current billing period"
    description: "Represents the status of a customer account"
  AccountLookupRequest:
    type: "object"
    required:
    - "accountRef"
    - "messageId"
    properties:
      accountRef:
        type: "string"
        description: "A reference number identifying the bill payments processor,\
          \ bill issuer, and customer"
        minLength: 6
        maxLength: 40
      messageId:
        $ref: "#/definitions/MessageId"
      merchant:
        $ref: "#/definitions/Merchant"
    description: "The data required to request account info"
  AccountLookupResponse:
    type: "object"
    required:
    - "account"
    - "receiver"
    properties:
      account:
        $ref: "#/definitions/Account"
      customer:
        $ref: "#/definitions/Customer"
      processor:
        $ref: "#/definitions/Institution"
      receiver:
        $ref: "#/definitions/Institution"
      slipData:
        $ref: "#/definitions/SlipData"
    description: "The data resulting from an account lookup request"
...
  Customer:
    type: "object"
    properties:
      firstName:
        type: "string"
        description: "The customer's first name(s)"
        minLength: 0
        maxLength: 40
      lastName:
        type: "string"
        description: "The customer's last name"
        minLength: 0
        maxLength: 40
      address:
        type: "string"
        description: "The customer's address"
        minLength: 0
        maxLength: 80
    description: "A bill payment customer"
  ErrorDetail:
    type: "object"
    required:
    - "responseMessage"
    properties:
      responseCode:
        type: "string"
        description: "The response code. Will always be present unless request to\
          \ upstream timed out"
      responseMessage:
        type: "string"
        description: "The response code description"
      detailMessage:
        type: "object"
        description: "A free form detailed description of a particular failure condition\
          \ may optionally be supplied"
    description: "Represents the outcome of a completed transaction"
  Institution:
    type: "object"
    required:
    - "id"
    - "name"
    properties:
      id:
        type: "string"
        description: "The institution's id"
        pattern: "[0-9]{1,11}"
      name:
        type: "string"
        description: "The institutions's name"
        minLength: 0
        maxLength: 40
    description: "Processing or receiving institution details"
  LedgerAmount:
    type: "object"
    required:
    - "amount"
    - "currency"
    properties:
      amount:
        type: "integer"
        format: "int64"
        description: "Amount in minor denomination, e.g. R799.95 is encoded as 79995"
      currency:
        type: "string"
        description: "Three digit currency number from ISO 4217, e.g. South African\
          \ Rand is encoded as 710"
        pattern: "[0-9]{0,12}"
      ledgerIndicator:
        type: "string"
        description: "Indicates whether this amount is a debit or a credit. Only required\
          \ when the amount can be either a debit or a credit"
        enum:
        - "DEBIT"
        - "CREDIT"
    description: "An amount object only containing value and currency, and optionally\
      \ an indicator of DEBIT/CREDIT"
  Merchant:
    type: "object"
    required:
    - "merchantId"
    - "merchantName"
    - "merchantType"
    properties:
      merchantType:
        type: "string"
        description: "The assigned four digit merchant category code"
        pattern: "[0-9]{4}"
      merchantId:
        type: "string"
        description: "The assigned merchant identifier. Also known as card acceptor\
          \ id"
        minLength: 15
        maxLength: 15
      merchantName:
        $ref: "#/definitions/MerchantName"
    description: "Merchant related data. Must be included if available"
  MerchantName:
    type: "object"
    required:
    - "city"
    - "country"
    - "name"
    - "region"
    properties:
      name:
        type: "string"
        description: "The merchant or trading as name associated with the merchant"
        minLength: 0
        maxLength: 23
      city:
        type: "string"
        description: "The city where the merchant is located"
        minLength: 0
        maxLength: 13
      region:
        type: "string"
        description: "The state or region where the merchant is located"
        minLength: 0
        maxLength: 2
      country:
        type: "string"
        description: "The country where the merchant is located"
        minLength: 0
        maxLength: 2
    description: "A container object representing the Cardholder Name and Location"
  MessageId:
    type: "object"
    required:
    - "counter"
    - "institutionId"
    - "senderId"
    - "time"
    properties:
      counter:
        type: "string"
        description: "An incrementing counter. Each seperate request must be issued\
          \ a new counter value. If the counter reaches 999999, it must wrap back\
          \ to 000000. Upon application startup, the starting value for the counter\
          \ should be randomly assigned rather than always starting from 000000. For\
          \ transactions initiated by a physical card entry or point-of-sale device,\
          \ this value is typically referred to as a System Trace Audit Number (STAN)\
          \ and there will be one counter per device. For transactions not initiated\
          \ by a physical device, each message should have a distinct counter value.\
          \ Typically this is achieved by assigning each entity sending transactions\
          \ a unique osenderId. Alternatively, effort must be taken by the client\
          \ to synchronize senders so that a unique senderId, time, and counter combination\
          \ is guaranteed. It is not necessary for messages to be delivered in the\
          \ same order as in which counter values where assigned"
        pattern: "[0-9]{6}"
      institutionId:
        type: "string"
        description: "The id of the instituation originating the request, as issued\
          \ by Electrum"
        pattern: "[0-9]{1,11}"
      senderId:
        type: "string"
        description: "The id that uniquely identifies each device or system in a sender\
          \ institution capabile of sending requests. For transactions initiated from\
          \ physical card entry or point-of-sale devices, this is the terminal id"
        pattern: "[a-zA-Z0-9]{8}"
      time:
        type: "string"
        format: "date-time"
        description: "The date and time of the request as recorded by the sender.\
          \ The format shall be as defined for date-time in [RFC 3339 section 5.6](https://tools.ietf.org/html/rfc3339#section-5.6).\
          \ It is recommended that the optional time-secfrac be included up to millisecond\
          \ precision"
    description: "A MessageId object uniquely identifies a request. It is very important\
      \ that this field is constructed carefully, since a mistake here may lead to\
      \ hard to find duplicate messages that typically only show up in a busy production\
      \ environment"
...
  SlipData:
    type: "object"
    required:
    - "issuerReference"
    properties:
      messageLines:
        type: "array"
        description: "An array of free text lines to be printed on the customer slip"
        items:
          type: "string"
      phoneNumber:
        type: "string"
        description: "The call centre phone number"
      issuerReference:
        type: "string"
        description: "An identifier that is printed on the customer slip and uniquely\
          \ identifies the payment on the bill issuer's system. This value is used\
          \ by the customer to request a refund when the service supports this function,\
          \ and it is thus important that this number is unique"
        pattern: "[A-Z0-9]{1,20}"
    description: "Data that may be printed on the custoemr slip for information purposes"
...